Unify

+++ goldendict.portable/tools/ChocolateyInstall.ps1
@@ -4,9 +4,11 @@ $PackageName = 'GoldenDict'
 $Url32 = 'https://sourceforge.net/projects/goldendict/files/early%20access%20builds/Qt5-based/GoldenDict-1.5.0-RC2-372-gc3ff15f_%28QT_5123%29.7z/download'
 $Url64 = 'https://sourceforge.net/projects/goldendict/files/early%20access%20builds/Qt5-based/64bit/GoldenDict-1.5.0-RC2-372-gc3ff15f_%28QT_5123%29%2864bit%29.7z/download'
 $InstallationPath = Join-Path $(Get-ToolsLocation) $PackageName
-$UnzipPath = Get-ToolsLocation
+$UnzipLocation = Join-Path $InstallationPath 'tmp'
 
-Get-ChildItem -Path $InstallationPath -Exclude "content", "portable" -ErrorAction Ignore | Remove-Item -Recurse -Force -ErrorAction Ignore
+if (Test-Path $InstallationPath) {
+    Get-ChildItem -Path $InstallationPath -Exclude "content", "portable" -ErrorAction Ignore | Remove-Item -Recurse -Force -ErrorAction Ignore
+}
 
 $PackageArgs = @{
     PackageName   = $PackageName
@@ -15,3 +17,7 @@ $PackageArgs = @{
     UnzipLocation = $UnzipPath
 }
 Install-ChocolateyZipPackage @PackageArgs
+
+$UnzipPath = (Get-ChildItem $UnzipLocation -Directory | Where-Object Name -Match "${PackageName}" | Select-Object -First 1).FullName
+Copy-Item -Path $(Join-Path $UnzipPath '*') -Destination $InstallationPath -Recurse -Force
+Remove-Item -Path $UnzipLocation -Recurse -Force -ErrorAction Ignore
